requirements.txt
pyaccsharedmemory
kafka-python